Fix Flaky tests (#226)
