+++ arrow-fx-coroutines/src/test/kotlin/arrow/fx/coroutines/ArrowFxSpec.kt
@@ -25,7 +25,7 @@ import kotlin.math.abs
  * So for `0..10` it will generate at most a `Stream` with `10` `Chunk`s of `10` elements.
  */
 abstract class StreamSpec(
-  iterations: Int = 100,
+  iterations: Int = 350,
   val depth: IntRange = 0..100,
   spec: StreamSpec.() -> Unit = {}
 ) : ArrowFxSpec(iterations) {
@@ -36,7 +36,7 @@ abstract class StreamSpec(
     range: IntRange = depth
   ): Arb<Pull<O, R>> =
     Arb.choice<Pull<O, R>>(
-      Arb.bind(Arb.stream(arbO), arbR) { s, r ->
+      Arb.bind(Arb.stream(arbO, range), arbR) { s, r ->
         s.asPull().map { r }
       },
       arbR.map { Pull.just(it) } as Arb<Pull<O, R>>,
@@ -69,7 +69,7 @@ abstract class StreamSpec(
  * Simple overwritten Kotest FreeSpec to reduce stress on tests.
  */
 abstract class ArrowFxSpec(
-  private val iterations: Int = 100,
+  private val iterations: Int = 350,
   spec: ArrowFxSpec.() -> Unit = {}
 ) : FreeSpec() {
